requirements.txt
PyQt5==5.15.9
Pillow==9.5.0
python-dotenv==1.0.1